refactor : split calendar event in monthly view
main_calendar.dart is to long to read;;
